Test: Separate concerns between cli and test
Pull references to structopt out of test so that cli is solely
responsible for parsing structopts. This creates an API that can be used
so that different front ends can be utilized with the underlying
library by separating the concerns of command line parsing and the
underlying test module.
