+++ src/bin/cli.rs
@@ -47,12 +47,51 @@ pub struct StatOptions {
     pub command: Vec<String>,
 }
 
+
+/// Configuration settings for running test
+#[derive(Debug, StructOpt)]
+pub struct TestOptions {
+    // Verbose flag, provides additional output
+    #[structopt(short = "v", long = "verbose", help = "provide additional output")]
+    pub verbose: bool,
+
+    // Should list runnable tests instead of performing them
+    #[structopt(
+        short = "l",
+        long = "list",
+        help = "list runnable tests instead of running"
+    )]
+    pub should_list: bool,
+
+    // Should format output as json
+    #[structopt(short = "j", long = "json", help = "format output as json")]
+    pub json: bool,
+
+    // A comma-seperated list of tests to skip
+    #[structopt(
+        short = "s",
+        long = "skip",
+        help = "a comma-seperated list of tests to skip",
+        default_value = ""
+    )]
+    pub to_skip: String,
+
+    // A comma-seperated list of tests to run
+    #[structopt(
+        short = "o",
+        long = "only",
+        help = "a comma-seperated list of tests to run",
+        default_value = ""
+    )]
+    pub to_run: String,
+}
+
 fn main() {
     let opt = Opt::from_args();
     match opt {
         Opt::Stat(x) => {
             run_stat(x.event, x.command);
         },
-        Opt::Test(x) => run_test(&x),
+        Opt::Test(x) => run_test(x.verbose, x.should_list, x.json, x.to_skip, x.to_run),
     }
 }
